refactoring: refactoring filter-cover
