+++ src/features/filter/types.ts
@@ -5,3 +5,24 @@ export interface IFilterValues {
   date: string;
   time: string[];
 }
+
+export interface FilteringProps {
+  items: {
+    categories?: boolean;
+    radius?: boolean;
+    sort?: boolean;
+    date?: boolean;
+    time?: boolean;
+    userCategories?: boolean;
+  };
+  setFilteres?: (date: IFilterValues) => void;
+  notFoundFilter?: never;
+}
+
+export interface NotFoundFilterProps {
+  items?: never;
+  setFilteres?: (date: IFilterValues) => void;
+  notFoundFilter: boolean;
+}
+
+export type FilterProps = FilteringProps | NotFoundFilterProps;
